fix reviewer's remarks

+++ ewm-service/src/main/java/ru/practicum/ewm/user/service/UserServiceImpl.java
@@ -5,8 +5,10 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.ewm.OffsetLimitPageable;
+import ru.practicum.ewm.event.repository.EventRepository;
 import ru.practicum.ewm.exception.ConflictException;
 import ru.practicum.ewm.exception.NotFoundException;
+import ru.practicum.ewm.request.repository.RequestRepository;
 import ru.practicum.ewm.user.model.User;
 import ru.practicum.ewm.user.repository.UserRepository;
 
@@ -16,6 +18,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
+    private final EventRepository eventRepository;
+    private final RequestRepository requestRepository;
 
 
     @Override
@@ -35,7 +39,14 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(Long userId) {
         User user = getById(userId);
-        userRepository.delete(user);
+        Pageable pageable = OffsetLimitPageable.of(0, 10);
+
+        if (eventRepository.findAllByInitiatorId(userId, pageable).isEmpty()
+                && requestRepository.findAllByRequester(user).isEmpty()) {
+            userRepository.delete(user);
+        } else {
+            throw new ConflictException("user can not be deleted", "");
+        }
     }
 
     @Override
